[REL] price_security to V11. [FIX] Separate Class in diferent py files. [FIX] Change views names. [FIX] Update README [FIX] Delete api.one on models [REM] Migration folder.
