+++ price_security/security/security.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
-    <data noupdate="1">
-
+<odoo noupdate="1">
 <!-- grupos y categoria de grupos-->
 
     <record model="ir.module.category" id="module_prices_management">
@@ -43,7 +41,4 @@
         <field name="groups" eval="[(4, ref('price_security.group_restrict_prices'))]"/>
         <field name="perm_read" eval="False"/>
     </record>
-
-    </data>
 </odoo>
-
