+++ price_security/views/account_views.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
-    <data>
+<odoo>
         <record id="view_payment_term_form" model="ir.ui.view">
             <field name="name">account.payment.term.form</field>
             <field name="model">account.payment.term</field>
@@ -22,6 +21,4 @@
                 </tree>
             </field>
         </record>
-        
-    </data>
 </odoo>
