+++ price_security/views/partner_views.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
-    <data>
+<odoo>
         <record id="view_partner_property_term_form" model="ir.ui.view">
             <field name="name">res.partner.account.property.form.inherit</field>
             <field name="model">res.partner</field>
@@ -23,5 +22,4 @@
                 </field>
             </field>
         </record>
-    </data>
 </odoo>
