+++ price_security/views/product_views.xml
