+++ price_security/views/res_users_views.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
-    <data>
+<odoo>
         <!-- Form -->
         <record id="view_price_security_users_form" model="ir.ui.view">
             <field name="name">price_security.users.form</field>
@@ -20,6 +19,4 @@
                 </notebook>
             </field>
         </record>
-        
-    </data>
 </odoo>
