+++ price_security/views/sale_order_views.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
-    <data>
+<odoo>
         <!-- sale.order -->
         <record id="view_price_security_order_form" model="ir.ui.view">
             <field name="name">price_security.order.form</field>
@@ -8,6 +7,7 @@
             <field name="inherit_id" ref="sale.view_order_form"/>
             <field name="groups_id" eval="[(6, 0, [ref('group_restrict_prices')])]"/>
             <field name="arch" type="xml">
+                
                 <xpath expr="//field[@name='order_line']/form" position="inside">
                     <field name="product_can_modify_prices" invisible="1"/>
                 </xpath>
@@ -50,8 +50,7 @@
                 <xpath expr="//field[@name='order_line']/form//field[@name='tax_id']" position="attributes">
                     <attribute name="invisible">1</attribute>
                 </xpath>
+            -->
             </field>
         </record>
-        
-    </data>
 </odoo>
